requirements.txt
anthropic>=0.7.0
openai>=1.1.0
python-dotenv>=1.0.0
gradio>=4.0.0
PyMuPDF>=1.23.0
ollama>=0.1.0
tenacity>=8.2.0
python-json-logger>=2.0.7